pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "drydock"
version = "0.3.0"
description = "Agent workspace orchestration CLI"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
drydock = "drydock.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"drydock.templates.devcontainer_drydock" = ["Dockerfile", "devcontainer.json"]
